+++ nextjs-dashboard/app/dashboard/search/page.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { DozerInfo } from '@/app/ui/search/DozerCard';
 import { useEffect, useState } from 'react';
 import SearchResults from '@/app/ui/search/SearchResults';
 import {
@@ -19,7 +18,11 @@ import {
 } from '@mui/material';
 import * as EmailValidator from 'email-validator';
 import { phone } from 'phone';
+import { convertDataToDozers } from './utils';
+import { DozerInfo } from './types';
+//import { staticData } from '@/app/lib/api';
 // import { Button } from '@/app/ui/button';
+
 
 export default function Page() {
   const [includeSmall, setIncludeSmall] = useState(false);
@@ -40,6 +43,7 @@ export default function Page() {
   //const [isLoading, setLoading] = useState(true)
 
   // fetch dozer data from api
+  // TODO: investigate fetching data at build time (XSS issues)
   useEffect(() => {
     fetch('/api/cat')
       .then((res) => res.json())
@@ -266,48 +270,6 @@ export default function Page() {
   ) => {
     setOpWeightSliderValue(newValue as number[]);
   };
-
-  function convertDataToDozers(data: any) {
-    const dozerInfos: DozerInfo[] = data.models.map((m: any) => {
-      const powerSpec = m.specs.find(
-        (s: any) =>
-          s.spec_name.toLowerCase().includes('power') &&
-          !s.spec_name.toLowerCase().includes('speed'),
-      );
-      const powerString: string =
-        powerSpec == undefined ? 'Unknown' : powerSpec.spec_value[0];
-      const powerNumber: number | undefined =
-        powerSpec == undefined
-          ? undefined
-          : Number(powerSpec.spec_value[0].split(' ')[0]);
-
-      const weightSpec = m.specs.find((s: any) =>
-        s.spec_name.toLowerCase().includes('operating weight'),
-      );
-      const weigthString: string =
-        weightSpec == undefined ? 'Unknown' : weightSpec.spec_value[0];
-      const weightNumber: number | undefined =
-        weightSpec == undefined
-          ? undefined
-          : Number(weightSpec.spec_value[0].split(' ')[0]);
-      const productCategory = m.productCategory.split('/')[2];
-
-      const dozerInfo: DozerInfo = {
-        category: productCategory,
-        engineHp: powerNumber,
-        engineHpString: powerString,
-        makeName: m.brand,
-        modelId: m.idmodel,
-        modelName: m.model_name,
-        operatingWeightString: weigthString,
-        operatingWeight: weightNumber,
-        imageUrl: m.image_url,
-      };
-      return dozerInfo;
-    });
-
-    return dozerInfos;
-  }
 
   function filterCriteria(dozer: DozerInfo): boolean {
     let meetsCategoryCriteria: boolean = false;
