+++ nextjs-dashboard/app/dashboard/search/types.ts
@@ -1,0 +1,11 @@
+export interface DozerInfo {
+  modelId: string;
+  makeName: string;
+  modelName: string;
+  category: string;
+  engineHp: number | undefined;
+  engineHpString: string;
+  operatingWeight: number | undefined;
+  operatingWeightString: string;
+  imageUrl: string;
+}
